MOAR comments and MOAR cleanup
