+++ app/src/lib/trampoline/trampoline-tokens.ts
@@ -2,10 +2,6 @@ import { uuid } from '../uuid'
 
 const trampolineTokens = new Set<string>()
 
-export function isValidTrampolineToken(token: string) {
-  return trampolineTokens.has(token)
-}
-
 function requestTrampolineToken() {
   const token = uuid()
   trampolineTokens.add(token)
@@ -16,6 +12,17 @@ function revokeTrampolineToken(token: string) {
   trampolineTokens.delete(token)
 }
 
+/** Checks if a given trampoline token is valid. */
+export function isValidTrampolineToken(token: string) {
+  return trampolineTokens.has(token)
+}
+
+/**
+ * Allows invoking a function with a short-lived trampoline token that will be
+ * revoked right after the function finishes.
+ *
+ * @param fn Function to invoke with the trampoline token.
+ */
 export async function withTrampolineToken<T>(
   fn: (token: string) => Promise<T>
 ): Promise<T> {
